Split test file into bestTime.js and test file
